requirements.txt
numpy>=1.20.0
Pillow>=10.2.0
matplotlib>=3.3.0
